Проведён рефакторинг функций, часть функций вынесена в util.js
